+++ gtg/CommonCrime/src/main/java/dk/sdu/mmmi/swe/gtg/commoncrime/ICrimeAction.java
@@ -5,7 +5,5 @@ import dk.sdu.mmmi.swe.gtg.common.data.entityparts.TexturePart;
 
 public interface ICrimeAction {
     public void commit(Entity entity);
-    public TexturePart getRobbedAtmTexture();
-    public TexturePart getAtmTexture();
 
 }
